pyproject.toml
[tool.poetry]
name = "fika"
version = "0.1.0"
description = ""
authors = ["one-zero-eight <https://t.me/one_zero_eight>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.13.1"
aiogram-dialog = "^2.2.0"
chatgpt-md-converter = "^0.1.2"
colorlog = "^6.8.2"
matplotlib = "^3.10.0"
openai = "^1.58.1"
pandas = "^2.2.3"
pre-commit = "^4.0.1"
pydantic = "^2.9.2"
pyyaml = "^6.0.2"
redis = "^5.1.0"
ruff = "^0.8.2"

[tool.ruff]
line-length = 120
lint.ignore = [
    # wildcard imports are fine
    "F403",
    "F405",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
